Inject dynamic and uniform buffer extensions in WebGPU engine
