+++ packages/dev/core/src/Engines/WebGPU/Extensions/index.ts
@@ -2,7 +2,6 @@ import "./engine.alpha";
 import "./engine.computeShader";
 import "./engine.cubeTexture";
 import "./engine.debugging";
-import "./engine.dynamicBuffer";
 import "./engine.dynamicTexture";
 import "./engine.externalTexture";
 import "./engine.multiRender";
@@ -13,5 +12,4 @@ import "./engine.renderTarget";
 import "./engine.renderTargetCube";
 import "./engine.textureSampler";
 import "./engine.storageBuffer";
-import "./engine.uniformBuffer";
 import "./engine.videoTexture";
